server/requirements.txt
aio-pika==8.0.3
aiormq==6.3.3
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
Django==3.1.5
django-cors-headers==3.6.0
django-health-check==3.16.5
django-prometheus==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==2.1.0
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==6.0.2
packaging==20.8
pamqp==3.1.0
Pillow==9.1.1
pluggy==0.13.1
prometheus-client==0.14.1
psycopg2-binary==2.9.1
py==1.10.0
py3-validate-email==1.0.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytz==2022.1
requests==2.26.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
supervisor==4.2.4
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
yarl==1.7.2
